refactored fields + forms out of admin
