+++ constance/utils.py
@@ -1,12 +1,10 @@
-# -*- coding: utf-8 -*-
-
 from __future__ import print_function
 from __future__ import unicode_literals
 
 from importlib import import_module
 
 from constance import settings, config
-from constance.admin import ConstanceForm
+from constance.forms import ConstanceForm
 
 
 def import_module_attr(path):
@@ -15,7 +13,10 @@ def import_module_attr(path):
 
 
 def get_constance_values():
-    # copied from constance.admin.ConstanceAdmin.changelist_view
+    """
+    Get dictionary of values from the backend
+    :return:
+    """
     default_initial = ((name, options[0])
                        for name, options in settings.CONFIG.items())
     # Then update the mapping with actually values from the backend
